refactor: 하단 버튼 공통 컴포넌트로 리팩토링
