+++ snapsplit/src/features/trip/createTrip/steps/Step1_CountrySearch/CountryList.tsx
@@ -18,7 +18,7 @@ const Country = ({
       <p className="text-body-2">{countryName}</p>
       <button
         onClick={onClick}
-        className={`px-3 py-1 rounded-[30px] outline-1 outline-offset-[-1px] text-body-2 
+        className={`px-3 py-1 rounded-[30px] outline-1 outline-offset-[-1px] cursor-pointer text-body-2 
           ${isSelected ? 'outline-grey-650 text-grey-650' : 'outline-grey-250 text-grey-450'}`}
       >
         선택
