+++ snapsplit/src/features/trip/createTrip/steps/Step3_AddMember/index.tsx
@@ -1,5 +1,5 @@
-import SearchBar from '@/shared/components/SearchBar';
-import { motion } from 'framer-motion';
+import BottomCTAButton from '@/shared/components/ui/BottomCTAButton';
+import SearchBar from '@/shared/components/ui/SearchBar';
 
 type AddMemberSectionProps = { onNext: () => void };
 
@@ -17,17 +17,7 @@ const AddMemberSection = ({ onNext }: AddMemberSectionProps) => {
         </div>
         <SearchBar placeholder="사용자코드를 입력해주세요" />
       </div>
-
-      <motion.button
-        onClick={onNext}
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        exit={{ opacity: 0, y: 20 }}
-        transition={{ duration: 0.3 }}
-        className="flex w-full p-3 bg-neutral-500 rounded-xl shadow-[0px_-2px_10px_0px_rgba(108,108,108,0.12)] justify-center items-center text-label-1 text-neutral-50"
-      >
-        다음으로
-      </motion.button>
+      <BottomCTAButton label="다음으로" onClick={onNext} />
     </div>
   );
 };
