+++ snapsplit/src/features/trip/createTrip/steps/Step4_InputTripName/index.tsx
@@ -1,7 +1,7 @@
-import SearchBar from '@/shared/components/SearchBar';
+import SearchBar from '@/shared/components/ui/SearchBar';
 import { SearchBarType } from '@trip/createTrip/type';
 import TripImageUploader from '@trip/createTrip/steps/Step4_InputTripName/TripImageUploader';
-import { motion } from 'framer-motion';
+import BottomCTAButton from '@/shared/components/ui/BottomCTAButton';
 
 type InputTripNameProps = { onNext: () => void };
 
@@ -19,17 +19,7 @@ const InputTripName = ({ onNext }: InputTripNameProps) => {
         <TripImageUploader />
         <SearchBar type={SearchBarType.WithoutIcon} placeholder="여행명을 입력해주세요" />
       </div>
-
-      <motion.button
-        onClick={onNext}
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        exit={{ opacity: 0, y: 20 }}
-        transition={{ duration: 0.3 }}
-        className="flex w-full p-3 bg-neutral-500 rounded-xl shadow-[0px_-2px_10px_0px_rgba(108,108,108,0.12)] justify-center items-center text-label-1 text-neutral-50"
-      >
-        다음으로
-      </motion.button>
+      <BottomCTAButton label="다음으로" onClick={onNext} />
     </div>
   );
 };
