+++ snapsplit/src/shared/components/ui/SearchBar.tsx
